Move EstimatorPubResult to results
